refactor template, default and action scripts
